Extract "keys" assertion to own file
